Adding a richer patient view example (#31)
Added an improved patient-view example that sends several infographic cards, and a card using the encounter prefetch template to show most recent encounter and number of encounters on record. Does not implement suggestions or links but this could be added to in later iterations for #36. Gained a good understanding of CDS hooks from a new dev. perspective.