Support ESM Paths
This is somewhat annoying, but ESM path resolution differs from Node path resolution.

For example if you have a folder called foo, and it has an index.js, then in node path imports you can just import "foo" and it will figure out that you are importing index.js.

ESM path resolution doesn't do that, relative module paths have to be... absolutely qualified.

Unfortunately, at this moment there isn't anything in the build that will actually _break_ if you don't do this properly. There are also places where you don't need to do it - such as if you are importing a type declaration (since that import gets throw away during the compilation phase).

It _is_ possible to set `experimental-specifier-resolution=node` on a consuming project to fix this, but we don't want to add any set up requirements.

Ultimately, we probably need to use a module compiler in addition to TypeScript, which is what I've been trying to avoid.

This is good enough for 0.1.0 though.