Connectathon enhancements (#81)
removed override reasons from the API guide, and isReccomended. Cleaned up the patient view service text guidance. Deployed patient view to a fly server, which launches the cardiac risk SMART app. Modified goodRx example to be triggered by order-sign. Changed the typing for a card from an array of overrideReasons to an array of codings to match the spec.